add "if __name__ == '__main__'" block
this makes the code suitable for use as a module in other pythons scripts.

The problem with calling a function openly in a document is that when a .py file is imported, any code in the global scope will be executed immediately. Any code in a if __name__ == '__main__': will not be executed when the file is imported as a module.