lift state to parent component
use a constructor for this, setting the state of all 9 squares to null and adding each to a n array.
modified Board to instruct each individual square about its value (X, O or null) from the array created by the constructor.
got rid of the constructor in Square since it is no longer needed :)